refactor: reorganize Bff types and introduce new chains and token types
