+++ packages/core-js/src/types/bff/token.ts
@@ -1,0 +1,9 @@
+import { Network } from "./chains";
+
+export type Token = {
+  tokenId: string;
+  tokenAddress: string;
+  network: Network;
+  whitelisted: boolean | undefined;
+  onRampEnabled: boolean | undefined;
+};
